+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The [](./providers/ht.sr.md) provider for [Sourcehut](https://sourcehut.org) has been added.
 - The `semver` and `sha256` modules are now optional in cases where they would never be invoked.
      - The related code will raise an error if it tries to find the modules but cannot.
+- The `rel.conflicts` field has been added.
+  The field should be a list of package names.
+  If any of the packages listed are installed when installing this package,
+  the installation will fail.
 
 ## v1.3.0
 - Packages are now evaluated inside of a sandbox, preventing package tables from having side effects.